- null check op selectedInputId bij invoer; - bezig met nieuw actieblok ActionFeatureType_AttributeNames_Rename om meerde invoerkolommen te kunnen mappen naar uitvoerkolommen; - indien uitvoertabel beschikbaar dan worden in het overzicht nu ook hiervan de attributen getoond;

+++ src/java/nl/b3p/datastorelinker/gui/stripes/InputAction.java
@@ -96,7 +96,9 @@ public class InputAction extends DefaultAction {
         EntityManager em = JpaUtilServlet.getThreadEntityManager();
         Session session = (Session)em.getDelegate();
 
-        session.delete(session.get(Inout.class, selectedInputId));
+        if (selectedInputId != null && selectedInputId > 0) {
+            session.delete(session.get(Inout.class, selectedInputId));
+        }
 
         return list();
     }
